[수정] DetailViewController 수정
 - 모델 단일화
 - 진입 유형 제거
